update networking, permissions and workloads packages
