+++ pkg/networking/route.go
@@ -4,13 +4,37 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/argoproj-labs/argocd-operator/pkg/mutation"
 	routev1 "github.com/openshift/api/route/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	cntrlClient "sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
+	"github.com/argoproj-labs/argocd-operator/pkg/mutation"
 )
+
+var routeAPIFound = false
+
+// IsRouteAPIAvailable returns true if the Route API is present.
+func IsRouteAPIAvailable() bool {
+	return routeAPIFound
+}
+
+// SetRouteAPIFound sets the value of routeAPIFound to provided input
+func SetRouteAPIFound(found bool) {
+	routeAPIFound = found
+}
+
+// verifyRouteAPI will verify that the Route API is present.
+func VerifyRouteAPI() error {
+	found, err := argoutil.VerifyAPI(routev1.GroupName, routev1.GroupVersion.Version)
+	if err != nil {
+		return err
+	}
+	routeAPIFound = found
+	return nil
+}
 
 // RouteRequest objects contain all the required information to produce a route object in return
 type RouteRequest struct {
