+++ pkg/workloads/openshift.go
@@ -1,8 +1,9 @@
 package workloads
 
 import (
-	util "github.com/argoproj-labs/argocd-operator/pkg/util"
-	template "github.com/openshift/api/template/v1"
+	templatev1 "github.com/openshift/api/template/v1"
+
+	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
 )
 
 var (
@@ -14,16 +15,17 @@ func IsTemplateAPIAvailable() bool {
 	return templateAPIFound
 }
 
+// SetTemplateAPIFound sets the value of prometheusAPIFound to provided input
+func SetTemplateAPIFound(found bool) {
+	templateAPIFound = found
+}
+
 // VerifyTemplateAPI will verify that the template API is present.
 func VerifyTemplateAPI() error {
-	found, err := util.VerifyAPI(template.SchemeGroupVersion.Group, template.SchemeGroupVersion.Version)
+	found, err := argoutil.VerifyAPI(templatev1.SchemeGroupVersion.Group, templatev1.SchemeGroupVersion.Version)
 	if err != nil {
 		return err
 	}
 	templateAPIFound = found
 	return nil
-}
-
-func SetTemplateAPIFound(val bool) {
-	templateAPIFound = val
 }
